refactor(googlechat): fold types.go into googlechat.go
Other notifier packages (slack, msteamsv2, pagerduty, etc.) keep types
and logic in a single file. Match that convention instead of a separate
types.go.
